% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FizzBuzz.R
\name{FizzBuzz}
\alias{FizzBuzz}
\title{A function that plays the fizz buzz game.}
\usage{
FizzBuzz(vector)
}
\arguments{
\item{vector}{a vector of values}
}
\value{
a vector of fizz buzz results
}
\description{
The user inputs a vector of poa=sitive integers. The fuction
returns a vector of values. It checks to see if the number is divisible
by 3, 5, or both 3 and five, and responds with
fizz buzz if divisible by 3 and 5, fizz if divisible by 3, buzz, if
divisible by 5, or just the nummber if not divisible by either 3 or 5.
}
\examples{
FizzBuzz(1:30)
FizzBuzz(c(2,5,3,15,8))
}
