% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flagstaff_Weather.R
\docType{data}
\name{Flagstaff_Weather}
\alias{Flagstaff_Weather}
\title{A time series of weather in Flagstaff, AZ.}
\format{
a data frame with 25,449 observations
\describe{
  \item{DATE}{The date of observation as a POSIX date format.}
  \item{PRCP}{Daily precipitation.}
  \item{SNOW}{Daily snowfall in inches.}
  \item{TMAX}{MAximum daily temperature in Fahrenheit.}
  \item{TMIN}{Minimum daily temperature in Fahrenheit.}
}
}
\source{
\url{www.ncdc.noaa.gov}
}
\usage{
Flagstaff_Weather
}
\description{
A time series of weather in Flagstaff, AZ.
}
\keyword{datasets}
